requirements.txt
Django==2.1.2
django-filter==2.0.0
django-modeltranslation==0.12.1
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
Pillow==4.3.0
psycopg2==2.7.1
celery==4.2.0
gunicorn==19.7.0
redis==2.10.5
elasticsearch==6.3.1
elasticsearch-dsl==6.1.0
factory-boy==2.11.1
